+++ lib/gloss/visitor.rb
@@ -4,17 +4,10 @@
   ##### See src/ to make changes
 
 module Gloss
-  module Utils
-    module_function
-    def with_file_header(str)
-      "#{Visitor::FILE_HEADER}\n\n#{str}"
-    end
-  end
   class Visitor
     FILE_HEADER = "  #{(if Config.frozen_string_literals
       "# frozen_string_literal: true\n"
     end)}\n  ##### This file was generated by Gloss; any changes made here will be overwritten.\n  ##### See #{Config.src_dir}/ to make changes"
-    include Utils
     attr_reader(:"tree")
     def initialize(tree_hash, type_checker = nil, on_new_file_referenced = nil)
       @on_new_file_referenced = on_new_file_referenced
@@ -24,10 +17,11 @@ module Gloss
       @current_scope = nil
       @tree = tree_hash
       @type_checker = type_checker
+      @past_module_function = false
     end
     def run()
       rb_output = visit_node(@tree)
-      with_file_header(rb_output)
+      Utils.with_file_header(rb_output)
     end
     def visit_node(node, scope = Scope.new)
       src = Source.new(@indent_level)
@@ -83,6 +77,7 @@ case node.[](:"type")
 .add(class_type)
           end)
         when "ModuleNode"
+          already_module_function = @past_module_function.dup
           module_name = visit_node(node.[](:"name"))
           src.write_ln("module #{module_name}")
           current_namespace = (if @current_scope
@@ -109,6 +104,7 @@ case node.[](:"type")
 .add(module_type)
           end)
           src.write_ln("end")
+          @past_module_function = false if @past_module_function && !already_module_function
         when "DefNode"
           args = render_args(node)
           receiver = (if node.[](:"receiver")
@@ -132,11 +128,7 @@ case node.[](:"type")
             nil
           end), location:           build_location(node))]
           method_definition = RBS::AST::Members::MethodDefinition.new(name:           node.[](:"name")
-.to_sym, kind:           (if receiver
-            :"singleton"
-          else
-            :"instance"
-          end), types: method_types, annotations: EMPTY_ARRAY, location:           build_location(node), comment:           node.[](:"comment"), overload: false)
+.to_sym, kind: @past_module_function ? :singleton_instance : receiver ? :singleton : :instance, types: method_types, annotations: EMPTY_ARRAY, location:           build_location(node), comment:           node.[](:"comment"), overload: false)
           (if @current_scope
             @current_scope.members
 .<<(method_definition)
@@ -198,9 +190,12 @@ EMPTY_ARRAY          }
           call = "#{obj}#{name}#{opening_delimiter}#{args}#{(if has_parens
             ")"
           end)}#{block}"
-          (if @on_new_file_referenced && name.==("require_relative")
-            @on_new_file_referenced.call(name, true)
-          end)
+          case name
+          when "require_relative"
+            @on_new_file_referenced.call(name, true) if @on_new_file_referenced
+          when "module_function"
+            @past_module_function = true
+          end
           src.write_ln(call)
         when "Block"
           args = render_args(node)
