+++ sig/yaml.rbs
@@ -1,0 +1,3 @@
+class Yaml
+  def safe_load: (String) -> untyped
+end
