+++ src/lib/gloss.gl
@@ -17,6 +17,7 @@ require "gloss/visitor"
 require "gloss/errors"
 require "gloss/logger"
 require "gloss/prog_loader"
+require "gloss/utils"
 
 is_ci = ENV.fetch("CI") { false }
 require "gls" unless is_ci # a bit of a hack for now
