+++ src/lib/gloss/logger.gl
@@ -1,8 +1,6 @@
 module Gloss
   def self.logger
-    if @logger
-      @logger
-    else
+    @logger ||= begin
       env_log_level = ENV.fetch("LOG_LEVEL") { "INFO" }
       real_log_level = {
         "UNKNOWN" => Logger::UNKNOWN,
@@ -15,12 +13,7 @@ module Gloss
         nil => nil,
         "" => nil
       }.fetch env_log_level
-      @logger = Logger.new(real_log_level ? STDOUT : nil)
-      formatter = Logger::Formatter.new
-      @logger.formatter = proc do |severity, datetime, progname, msg|
-        formatter.call(severity, datetime, progname, msg)
-      end
-      @logger
+      Logger.new(real_log_level ? STDOUT : IO::NULL)
     end
   end
 end
