+++ src/lib/gloss/prog_loader.gl
@@ -1,45 +1,18 @@
 require "rubygems/gem_runner"
 
 module Gloss
-  module Utils
-    module_function
-
-    def absolute_path(path)
-      pn = Pathname.new(path)
-      if pn.absolute?
-        pn.to_s
-      else
-        ap = File.absolute_path path
-        if File.exist? ap
-          ap
-        else
-          throw :error, "File path #{path} does not exist (also looked for #{ap})"
-        end
-      end
-    end
-
-    def gem_path_for(gem_name)
-      Gem.ui.instance_variable_set :"@outs", StringIO.new
-      Gem::GemRunner.new.run(["which", gem_name])
-      Gem.ui.outs.string
-    rescue SystemExit => e
-      nil
-    end
-  end
-
   OUTPUT_BY_PATH = Hash.new
 
   class ProgLoader
-    include Utils
-
     def initialize
       entrypoint = Config.entrypoint
-      if entrypoint.nil? || entrypoint == ""
+      if entrypoint == nil || entrypoint == ""
         throw :error, "Entrypoint is not yet set in .gloss.yml"
       end
       @files_to_process = [
-        absolute_path(Config.entrypoint),
-        absolute_path(File.join(__dir__, "..", "..", "sig", "core.rbs"))
+        Utils.absolute_path(Config.entrypoint),
+        # __dir__ is typed as String? - but it shouldn't be nil here
+        Utils.absolute_path(File.join((__dir__||""), "..", "..", "sig", "core.rbs"))
       ]
       @processed_files = Set.new
       @type_checker = TypeChecker.new
@@ -47,28 +20,22 @@ module Gloss
 
     def run
       @files_to_process.each do |path_string|
-        next if @processed_files.member?(path_string) || OUTPUT_BY_PATH.fetch(path_string) { nil }
-
-        Gloss.logger.debug "Loading #{path_string}"
-        path = absolute_path(path_string)
-        file_contents = File.open(path).read
-        contents_tree = Parser.new(file_contents).run
-        on_new_file_referenced = proc do |pa, relative|
-          pn = if relative
-                 require_relative_path_to_file_path pa
-               else
-                 require_path_to_file_path pa
-               end
-          if pn
-            @files_to_process << pn
-          elsif Config.strict_require
-            throw :error, "Cannot resolve require path for #{pa}"
-          else
-            # no op
+        # currently steep would give an `unexpected jump` if next was used
+        unless @processed_files.member?(path_string) || OUTPUT_BY_PATH.[](path_string)
+          Gloss.logger.debug "Loading #{path_string}"
+          path = Utils.absolute_path(path_string)
+          file_contents = File.open(path).read
+          contents_tree = Parser.new(file_contents).run
+          on_new_file_referenced = proc do |pa, relative|
+            if relative
+              handle_require_relative pa
+            else
+              handle_require pa
+            end
           end
+          OUTPUT_BY_PATH.[](path_string) = Visitor.new(contents_tree, @type_checker, on_new_file_referenced).run
+          @processed_files.add path_string
         end
-        OUTPUT_BY_PATH.[](path_string) = Visitor.new(contents_tree, @type_checker, on_new_file_referenced).run
-        @processed_files.add path_string
       end
 
       @type_checker
@@ -99,7 +66,7 @@ module Gloss
         # no .gl file available - .rbs file available?
         # TODO: verify file is still actually requireable
         pathn = Pathname.new("#{File.join(Dir.pwd, "sig", path)}.rbs")
-        gem_path = gem_path_for(path)
+        gem_path = Utils.gem_path_for(path)
         if gem_path
           sig_files = Dir.glob(File.absolute_path(File.join(gem_path, "..", "..", "sig", "**", "*.rbs")))
           if sig_files.length.positive?
@@ -127,7 +94,7 @@ module Gloss
               rbs_type_deps.each { |d| load_rbs_from_require_path d }
             end
           elsif Config.strict_require
-            throw :error, "Cannot resolve require path for #{pa}"
+            throw :error, "Cannot resolve require path for #{path}"
           else
             Gloss.logger.debug "No path found for #{path}"
           end
@@ -137,7 +104,7 @@ module Gloss
 
     private def handle_require_relative(path)
       base = File.join(@filepath, "..", path)
-      pn = nil
+      pn : String? = nil
       Gem.suffixes.each do |ext|
         full = File.absolute_path(base + ext)
         pn = full if File.exist?(full)
@@ -146,7 +113,7 @@ module Gloss
       if pn
         @files_to_process << pn unless @files_to_process.include? pn
       elsif Config.strict_require
-        throw :error, "Cannot resolve require path for #{pa}"
+        throw :error, "Cannot resolve require path for #{pn}"
       else
         Gloss.logger.debug "No path found for #{path}"
       end
