+++ src/lib/gloss/visitor.gl
@@ -1,20 +1,10 @@
 module Gloss
-  module Utils
-    module_function
-
-    def with_file_header(str)
-      "#{Visitor::FILE_HEADER}\n\n#{str}"
-    end
-  end
-
   class Visitor
     FILE_HEADER = <<~RUBY
       #{"# frozen_string_literal: true\n" if Config.frozen_string_literals}
       ##### This file was generated by Gloss; any changes made here will be overwritten.
       ##### See #{Config.src_dir}/ to make changes
     RUBY
-
-    include Utils
 
     attr_reader :tree
 
@@ -25,11 +15,12 @@ module Gloss
       @current_scope = nil
       @tree = tree_hash
       @type_checker = type_checker
+      @past_module_function = false
     end
 
     def run
       rb_output = visit_node(@tree)
-      with_file_header(rb_output)
+      Utils.with_file_header(rb_output)
     end
 
     # type node = Hash[Symbol, String | Array[String | node] | Hash[Symbol, node]] | true | false
@@ -94,6 +85,7 @@ module Gloss
           @type_checker.top_level_decls.add(class_type)
         end
       when "ModuleNode"
+        already_module_function = @past_module_function.dup
         module_name = visit_node node[:name]
         src.write_ln "module #{module_name}"
 
@@ -124,6 +116,7 @@ module Gloss
           @type_checker.top_level_decls.add(module_type)
         end
         src.write_ln "end"
+        @past_module_function = false if @past_module_function && !already_module_function
       when "DefNode"
         args = render_args(node)
         receiver = node[:receiver] ? visit_node(node[:receiver]) : nil
@@ -176,7 +169,7 @@ module Gloss
         ]
         method_definition = RBS::AST::Members::MethodDefinition.new(
           name: node[:name].to_sym,
-          kind: receiver ? :singleton : :instance,
+          kind: @past_module_function ? :singleton_instance : receiver ? :singleton : :instance,
           types: method_types,
           annotations: EMPTY_ARRAY, # TODO
           location: build_location(node),
@@ -217,7 +210,12 @@ module Gloss
                             end
         name = node[:name]
         call = "#{obj}#{name}#{opening_delimiter}#{args}#{")" if has_parens}#{block}"
-        @on_new_file_referenced.call(name, true) if @on_new_file_referenced && name == "require_relative"
+        case name
+        when "require_relative"
+          @on_new_file_referenced.call(name, true) if @on_new_file_referenced
+        when "module_function"
+          @past_module_function = true
+        end
         src.write_ln(call)
 
       when "Block"
