+++ src/lib/gloss/writer.gl
@@ -4,22 +4,13 @@ require "pathname"
 require "fileutils"
 
 module Gloss
-  module Utils
-    module_function
-
-    def src_path_to_output_path(src_path : String) : String
-      src_path.sub("#{Config.src_dir}/", "")
-              .sub(/\.gl$/, ".rb")
-    end
-  end
-
   class Writer
-    include Utils
-
     def initialize(
       @content,
       @src_path : String,
-      @output_path : Pathname? = Pathname.new(src_path_to_output_path(src_path))
+      @output_path : Pathname? = Pathname.new(
+        Utils.src_path_to_output_path(src_path)
+      )
     )
     end
